Moved functions to appropriate modules and refactored some function names
